Move ValidMethod out of plugin
