+++ validator/pom.xml
@@ -24,7 +24,15 @@
       <artifactId>avaje-validator-http-plugin</artifactId>
       <version>${project.version}</version>
     </dependency>
-
+    
+    <dependency>
+      <groupId>io.avaje</groupId>
+      <artifactId>avaje-inject</artifactId>
+      <version>${inject.version}</version>
+      <scope>provided</scope>
+      <optional>true</optional>
+    </dependency>
+    
     <dependency>
       <groupId>jakarta.validation</groupId>
       <artifactId>jakarta.validation-api</artifactId>
