+++ validator/src/main/java/io/avaje/validation/ValidMethod.java
@@ -1,4 +1,4 @@
-package io.avaje.validation.inject.aspect;
+package io.avaje.validation;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
@@ -8,12 +8,11 @@ import java.lang.annotation.Target;
 
 import io.avaje.inject.aop.Aspect;
 
-@Deprecated
 @Aspect
 @Target(METHOD)
 @Retention(RUNTIME)
 /** Place on a method to execute validations on the parameters and return types */
-public @interface ValidateMethod {
+public @interface ValidMethod {
   String locale() default "";
 
   boolean throwOnParamFailure() default false;
